Add files via upload
1. Reorganize folders for projects.
2. Add reflection and panel of experts (POE).  

Definitions:
Reflection means this: A request to an LLM is made, the response is sent back to the LLM with a request for it to try to improve it.  
POE means this: Multiple LLMs are sent same request. Then, all the responses are numbered, and sent to each LLM asking it to select which response is superior, by indicating its number.